feat: add font

+++ src/textui/textmode.h
@@ -154,7 +154,7 @@ void textmode_dblrect(
 	uint8_t color
 );
 
-void textmode_init_font(const uint8_t offset, const uint8_t count, const uint8_t charHeight, const uint8_t *charData);
+void textmode_init_font(const uint16_t offset, const uint16_t count, const uint16_t charHeight, const uint8_t *charData);
 
 void textmode_gotoxy(const uint8_t x, const uint8_t y);
